Renamed file for central node
